feat: modularized project structure, updating make for the changes
