+++ .gitignore
@@ -1,1 +1,2 @@
 cube
+build/
